requirements_project.txt
# Requirements file generated by analyzing all Python files in the project
# This includes all dependencies found in the non-ignored Python files

# Core vector database and client libraries
qdrant_client

# API framework and related packages
fastapi[all]
pydantic

# AI/ML and embedding libraries
openai

# Environment and configuration management
python-dotenv

# HTTP requests
requests

# Terminal output styling and colors
colorama

# Data visualization and plotting
matplotlib

# Testing framework (HyperSkill specific)
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz

# Standard library modules used (no installation required):
# - json (built-in)
# - os (built-in) 
# - time (built-in)
# - uuid (built-in)
# - typing (built-in)
# - re (built-in)
# - itertools (built-in)
# - statistics (built-in)
# - pprint (built-in)